separation of interfaces from the corresponding classes to follow in place convention, also remove of the last of the dependencies of the DataAccess layer from the Campaign controller, this included the creation of MemberSearchFilterService layer
