+++ src/GiveCRM.Web/GiveCRM.Web.csproj
@@ -110,10 +110,13 @@
     <Compile Include="Services\CampaignService.cs" />
     <Compile Include="Services\DonationsService.cs" />
     <Compile Include="Services\FacetsService.cs" />
+    <Compile Include="Services\ICampaignService.cs" />
     <Compile Include="Services\IDonationsService.cs" />
     <Compile Include="Services\IFacetsService.cs" />
     <Compile Include="Services\IMailingListService.cs" />
+    <Compile Include="Services\IMemberSearchFilterService.cs" />
     <Compile Include="Services\IMemberService.cs" />
+    <Compile Include="Services\ISearchService.cs" />
     <Compile Include="Services\MailingListService.cs" />
     <Compile Include="Services\MemberService.cs" />
     <Compile Include="Services\SearchService.cs" />
