+++ src/GiveCRM.Web/Services/CampaignService.cs
@@ -34,13 +34,4 @@ namespace GiveCRM.Web.Services
         }
 
     }
-
-    public interface ICampaignService
-    {
-        IEnumerable<Campaign> AllOpen();
-        IEnumerable<Campaign> AllClosed();
-        Campaign Get(int id);
-        Campaign Insert(Campaign campaign);
-        void Update(Campaign campaign);
-    }
 }
